# perf(core): optimize encode on large strings