Follow up to serde_v8's #14450

```
Before:
test bench_utf8_decode_12_b  ... bench:         188 ns/iter (+/- 41)
test bench_utf8_decode_12_kb ... bench:       3,480 ns/iter (+/- 2,226)
test bench_utf8_decode_12_mb ... bench:   1,549,520 ns/iter (+/- 210,202)
test bench_utf8_encode_12_b  ... bench:         400 ns/iter (+/- 69)
test bench_utf8_encode_12_kb ... bench:       5,813 ns/iter (+/- 469)
test bench_utf8_encode_12_mb ... bench:   5,149,035 ns/iter (+/- 160,745)

After:
test bench_utf8_decode_12_b  ... bench:         174 ns/iter (+/- 13)
test bench_utf8_decode_12_kb ... bench:       3,780 ns/iter (+/- 688)
test bench_utf8_decode_12_mb ... bench:   1,484,499 ns/iter (+/- 42,377)
test bench_utf8_encode_12_b  ... bench:         345 ns/iter (+/- 22)
test bench_utf8_encode_12_kb ... bench:       2,127 ns/iter (+/- 132)
test bench_utf8_encode_12_mb ... bench:   1,198,063 ns/iter (+/- 83,304)
```

We observe non-linear improvements as input string length grows, approaching ~3x for multi-kb strings and beyond